Refactor namespaces to unify event sourcing under Sekiban.Pure.Events
