+++ samples/AspireEventSample/AspireEventSample.ApiService/Grains/IBranchEntityWriter.cs
@@ -1,4 +1,4 @@
-using Sekiban.Pure.OrleansEventSourcing;
+using Sekiban.Pure.Orleans;
 
 namespace AspireEventSample.ApiService.Grains;
