+++ samples/AspireEventSample/AspireEventSample.ApiService/Program.cs
@@ -10,7 +10,7 @@ using Sekiban.Pure.AspNetCore;
 using Sekiban.Pure.Command.Handlers;
 using Sekiban.Pure.CosmosDb;
 using Sekiban.Pure.Documents;
-using Sekiban.Pure.OrleansEventSourcing;
+using Sekiban.Pure.Orleans;
 using Sekiban.Pure.Postgres;
 using Sekiban.Pure.Projectors;
 var builder = WebApplication.CreateBuilder(args);
