+++ samples/AspireEventSample/Sekiban.Pure.Orleans.NUnit/Sekiban.Pure.Orleans.NUnit.csproj
@@ -1,4 +1,4 @@
-﻿﻿<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <TargetFramework>net9.0</TargetFramework>
@@ -8,7 +8,7 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0"/>
-        <PackageReference Include="Microsoft.Orleans.TestingHost" Version="8.0.0"/>
+        <PackageReference Include="Microsoft.Orleans.TestingHost" Version="9.1.2" />
         <PackageReference Include="NUnit" Version="4.3.2"/>
     </ItemGroup>
