+++ samples/AspireEventSample/Sekiban.Pure.Orleans/AggregateEventHandlerGrain.cs
@@ -1,7 +1,7 @@
 using ResultBoxes;
 using Sekiban.Pure.Documents;
 using Sekiban.Pure.Events;
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 public class AggregateEventHandlerGrain(
     [PersistentState("aggregate", "Default")]
