+++ samples/AspireEventSample/Sekiban.Pure.Orleans/IAggregateProjectorGrain.cs
@@ -1,6 +1,6 @@
 using Sekiban.Pure.Command.Handlers;
 
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 public interface IAggregateProjectorGrain : IGrainWithStringKey
 {
