+++ samples/AspireEventSample/Sekiban.Pure.Orleans/IEntityWriter.cs
@@ -1,4 +1,4 @@
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 public interface IEntityWriter<TEntity> where TEntity : IReadModelEntity
 {
