+++ samples/AspireEventSample/Sekiban.Pure.Orleans/IMultiProjectorGrain.cs
@@ -1,6 +1,6 @@
 using Sekiban.Pure.Query;
 
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 public interface IMultiProjectorGrain : IGrainWithStringKey
 {
