+++ samples/AspireEventSample/Sekiban.Pure.Orleans/IOrleansQueryResult.cs
@@ -1,3 +1,3 @@
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 public interface IOrleansQueryResult;
