+++ samples/AspireEventSample/Sekiban.Pure.Orleans/OrleansCommandResponse.cs
@@ -1,4 +1,4 @@
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 [GenerateSerializer]
 public record OrleansCommandResponse(
