+++ samples/AspireEventSample/Sekiban.Pure.Orleans/OrleansEmptyAggregatePayload.cs
@@ -1,6 +1,6 @@
 using Sekiban.Pure.Aggregates;
 
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 [GenerateSerializer]
 public record OrleansEmptyAggregatePayload() : IAggregatePayload;
