+++ samples/AspireEventSample/Sekiban.Pure.Orleans/OrleansEventExtensions.cs
@@ -2,7 +2,7 @@ using ResultBoxes;
 using Sekiban.Pure.Documents;
 using Sekiban.Pure.Events;
 
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 public static class OrleansEventExtensions
 {
