+++ samples/AspireEventSample/Sekiban.Pure.Orleans/OrleansEventMetadata.cs
@@ -1,5 +1,5 @@
 using Sekiban.Pure.Events;
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 [GenerateSerializer]
 public record OrleansEventMetadata([property:Id(0)]string CausationId,
