+++ samples/AspireEventSample/Sekiban.Pure.Orleans/OrleansListQueryResultGeneral.cs
@@ -1,5 +1,5 @@
 using Sekiban.Pure.Query;
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 [GenerateSerializer]
 public record OrleansListQueryResultGeneral(
