+++ samples/AspireEventSample/Sekiban.Pure.Orleans/OrleansQueryResultGeneral.cs
@@ -1,6 +1,6 @@
 using Sekiban.Pure.Query;
 
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 [GenerateSerializer]
 public record OrleansQueryResultGeneral(
