+++ samples/AspireEventSample/Sekiban.Pure.Orleans/SekibanOrleansExecutor.cs
@@ -7,7 +7,7 @@ using Sekiban.Pure.Events;
 using Sekiban.Pure.Executors;
 using Sekiban.Pure.Projectors;
 using Sekiban.Pure.Query;
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Orleans;
 
 public class SekibanOrleansExecutor(
     IClusterClient clusterClient,
