+++ samples/AspireEventSample/Sekiban.Pure.Postgres/SekibanPostgresExtensions.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Sekiban.Pure.OrleansEventSourcing;
+using Sekiban.Pure.Events;
 namespace Sekiban.Pure.Postgres;
 
 public static class SekibanPostgresExtensions
