+++ samples/AspireEventSample/Sekiban.Pure/Events/IEventReader.cs
@@ -1,6 +1,5 @@
 using ResultBoxes;
-using Sekiban.Pure.Events;
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Events;
 
 public interface IEventReader
 {
