+++ samples/AspireEventSample/Sekiban.Pure/Events/ISortableIdCondition.cs
@@ -1,6 +1,6 @@
 using Sekiban.Pure.Aggregates;
 using Sekiban.Pure.Documents;
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Events;
 
 public interface ISortableIdCondition
 {
