+++ samples/AspireEventSample/Sekiban.Pure/Events/SortableIdConditionNone.cs
@@ -1,5 +1,5 @@
 using Sekiban.Pure.Documents;
-namespace Sekiban.Pure.OrleansEventSourcing;
+namespace Sekiban.Pure.Events;
 
 public record SortableIdConditionNone : ISortableIdCondition
 {
