+++ samples/AspireEventSample/Sekiban.Pure/Sekiban.Pure.csproj
@@ -13,8 +13,4 @@
         <PackageReference Include="ResultBoxes" Version="0.3.30"/>
     </ItemGroup>
 
-    <ItemGroup>
-        <Folder Include="Cache\"/>
-    </ItemGroup>
-
 </Project>
